+++ tests/test_validators.py
@@ -7,8 +7,9 @@ from __future__ import absolute_import, division, print_function
 import pytest
 import zope.interface
 
-from attr.validators import instance_of, provides, optional
+from attr.validators import and_, instance_of, provides, optional
 from attr._compat import TYPE
+from attr._make import attributes, attr
 
 from .utils import simple_attr
 
@@ -56,6 +57,53 @@ class TestInstanceOf(object):
             "<instance_of validator for type <{type} 'int'>>"
             .format(type=TYPE)
         ) == repr(v)
+
+
+def always_pass(_, __, ___):
+    """
+    Toy validator that always passses.
+    """
+
+
+def always_fail(_, __, ___):
+    """
+    Toy validator that always fails.
+    """
+    0/0
+
+
+class TestAnd(object):
+    def test_success(self):
+        """
+        Succeeds if all wrapped validators succeed.
+        """
+        v = and_(instance_of(int), always_pass)
+
+        v(None, simple_attr("test"), 42)
+
+    def test_fail(self):
+        """
+        Fails if any wrapped validator fails.
+        """
+        v = and_(instance_of(int), always_fail)
+
+        with pytest.raises(ZeroDivisionError):
+            v(None, simple_attr("test"), 42)
+
+    def test_sugar(self):
+        """
+        `and_(v1, v2, v3)` and `[v1, v2, v3]` are equivalent.
+        """
+        @attributes
+        class C(object):
+            a1 = attr("a1", validator=and_(
+                instance_of(int),
+            ))
+            a2 = attr("a2", validator=[
+                instance_of(int),
+            ])
+
+        assert C.__attrs_attrs__[0].validator == C.__attrs_attrs__[1].validator
 
 
 class IFoo(zope.interface.Interface):
@@ -111,12 +159,6 @@ class TestProvides(object):
         ) == repr(v)
 
 
-def always_pass(_, __, ___):
-    """
-    Toy validator that always passses.
-    """
-
-
 @pytest.mark.parametrize("validator", [
     instance_of(int),
     [always_pass, instance_of(int)],
@@ -162,13 +204,13 @@ class TestOptional(object):
 
         if isinstance(validator, list):
             assert (
-                ("<optional validator for ({func}, <instance_of validator for "
-                 "type <{type} 'int'>>) or None>")
+                ("<optional validator for _AndValidator(_validators=[{func}, "
+                 "<instance_of validator for type <{type} 'int'>>]) or None>")
                 .format(func=repr(always_pass), type=TYPE)
             ) == repr(v)
         else:
             assert (
-                ("<optional validator for (<instance_of validator for type "
-                 "<{type} 'int'>>,) or None>")
+                ("<optional validator for <instance_of validator for type "
+                 "<{type} 'int'>> or None>")
                 .format(type=TYPE)
             ) == repr(v)