+++ CHANGELOG.rst
@@ -52,7 +52,9 @@ Changes:
   `#128 <https://github.com/python-attrs/attrs/pull/128>`_
 - ``__attrs_post_init__()`` is now run if validation is disabled.
   `#130 <https://github.com/python-attrs/attrs/pull/130>`_
-- The ``validator`` argument of ``@attr.s`` now can take a ``list`` of validators that all have to pass.
+- Added ``attr.validators.and_()`` that composes multiple validators into one.
+  `#161 <https://github.com/python-attrs/attrs/issues/161>`_
+- For convenience, the ``validator`` argument of ``@attr.s`` now can take a ``list`` of validators that are wrapped using ``and_()``.
   `#138 <https://github.com/python-attrs/attrs/issues/138>`_
 - Accordingly, ``attr.validators.optional()`` now can take a ``list`` of validators too.
   `#161 <https://github.com/python-attrs/attrs/issues/161>`_